requirements.txt
tinygrad
albumentations
numpy
pillow
pytest
PyYAML
